Users not logged in will be redirected to log in
In our prod app, we would cache the 'Write a review' piece of the page when someone was logged in. Now someone who is logged out could see it, but they would get a blank page b/c of the authorize! call. Now they get redirected with a friendly message.